Mostly Complete!
This updates basically covered everything that needed to be done past the generation of a solution at the beginning. Once a board with a full solution is created, the program will now make a number of cells 0, allow the user to edit those cells in a certain way, give the user an incomplete board and prompt them to fill it in. Once the user thinks that they have solved the puzzle correctly, the game checks whether the solution the use gave is valid and informs them. Minor details remain, mostly documentation related.